requirements.txt
Librerías necesarias: 
streamlit
torch
transformers
Pillow